% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_drender}
\alias{addin_drender}
\title{RStudio Addin: Render the Dockerized R Markdown Document}
\usage{
addin_drender()
}
\description{
Call this function as an addin to render (drender) the current document.
}

