% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_lift_drender}
\alias{addin_lift_drender}
\title{RStudio Addin: Dockerize and Render the R Markdown Document}
\usage{
addin_lift_drender()
}
\description{
Call this function as an addin to dockerize (lift) and render (drender)
the current document.
}

