% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_lift_render_docker}
\alias{addin_lift_render_docker}
\title{RStudio Addin: Containerize and Render R Markdown Document with Docker}
\usage{
addin_lift_render_docker()
}
\description{
Call this function as an addin to containerize and render
the current document.
}

