% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_new_drmd}
\alias{addin_new_drmd}
\title{RStudio Addin: New R Markdown Document with Docker (and Rabix) Support}
\usage{
addin_new_drmd()
}
\description{
Call this function as an addin to create new R Markdown document with
Docker (and Rabix) support.
}

