% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_prune_dangling}
\alias{addin_prune_dangling}
\title{RStudio Addin: Remove Dangling Container + Image}
\usage{
addin_prune_dangling()
}
\description{
RStudio addin for removing any dangling Docker
containers and images automatically.
}
\keyword{internal}
