% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_prune_image}
\alias{addin_prune_image}
\title{RStudio Addin: Remove the Built Docker Image}
\usage{
addin_prune_image()
}
\description{
RStudio addin for removing the Docker images used for
rendering the current document.
}
\keyword{internal}
