% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_docker_install}
\alias{check_docker_install}
\title{Check if Docker was Installed}
\usage{
check_docker_install()
}
\value{
\code{TRUE} if Docker was deteted, \code{FALSE} otherwise.
}
\description{
This function checks if Docker was properly
installed and discoverable by R and liftr.
If still not usable, please start Docker daemon
}
\examples{
\dontrun{
check_docker_install()}
}
