% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{check_docker_running}
\alias{check_docker_running}
\title{Check if Docker Daemon is Running}
\usage{
check_docker_running()
}
\value{
\code{TRUE} if Docker daemon is running, \code{FALSE} otherwise.
}
\description{
This function checks if the Docker daemon is running.
}
\examples{
\dontrun{
check_docker_running()}
}
