% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drender.R
\name{drender}
\alias{drender}
\title{Render Dockerized R Markdown Documents}
\usage{
drender(input = NULL, tag = NULL, build_args = NULL,
  container_name = NULL, no_cache = TRUE, ...)
}
\arguments{
\item{input}{Input file to render in Docker container.}

\item{tag}{Docker image name to build, sent as docker argument \code{-t}.
If not specified, it will use the same name as the input file.}

\item{build_args}{A character string specifying additional
\code{docker build} arguments. For example,
\code{--pull=true -m="1024m" --memory-swap="-1"}.}

\item{container_name}{Docker container name to run.
If not specified, we will generate and use a random name.}

\item{no_cache}{Logical. Controls the \code{--no-cache} argument
in \code{docker run}. Setting this to be \code{TRUE} can accelerate
the rendering speed substantially for repeated compilation since
most of the Docker image layers will be cached, with only the
changed (knitr related) image layer being updated. Default is \code{TRUE}.}

\item{...}{Additional arguments passed to
\code{\link[rmarkdown]{render}}.}
}
\value{
Rendered file is written to the same directory of the input file.
A character vector with the image name and container name will be
returned. You will be able to manage them with \code{docker}
commands later or with the cleanup functions.
}
\description{
Render dockerized R Markdown documents using Docker containers.
}
\details{
Before using \code{drender()}, run \link{lift} on the document
first to generate \code{Dockerfile}.
See \code{vignette('liftr-intro')} for details about the extended
YAML front-matter metadata format and system requirements for
rendering dockerized R Markdown documents.
}
\examples{
# Dockerized R Markdown document
# Docker is required to run the example,
# so make sure we can use `docker` in terminal.
dir_docker = paste0(tempdir(), '/drender_docker/')
dir.create(dir_docker)
file.copy(system.file("docker.Rmd", package = "liftr"), dir_docker)
docker_input = paste0(dir_docker, "docker.Rmd")
lift(docker_input)
\dontrun{
drender(docker_input)
# view rendered document
browseURL(paste0(dir_docker, "docker.html"))}
}

