% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_docker}
\alias{install_docker}
\title{Installation Helper for Docker Engine}
\usage{
install_docker()
}
\description{
This function guides you to install Docker (Engine).
}
\examples{
\dontrun{
install_docker()}
}
\references{
\url{https://docs.docker.com/engine/installation/}
}
