% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{install_from_rmd}
\alias{install_from_rmd}
\title{install from liftr rmarkdown with headers}
\usage{
install_from_rmd(rmd)
}
\arguments{
\item{rmd}{A rmarkdown with lift header}
}
\description{
install from liftr rmarkdown with headers
}
\examples{
\dontrun{
install_from_rmd("test.rmd")
}
}

