% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{lift_docopt}
\alias{lift_docopt}
\title{lift a docopt string}
\usage{
lift_docopt(input)
}
\arguments{
\item{input}{input Rmarkdown file or a function name (character)}
}
\value{
a string used for docopt
}
\description{
lift a docopt string used for command line
}
\details{
parse Rmarkdown header from rabix field
}
\examples{
fl = system.file("examples/runif.Rmd", package = "liftr")
opts = lift_docopt(fl)
\dontrun{
require(docopt)
docopt(opts)
docopt(lift_docopt("mean.default"))
}
}

