% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{lift_shinyapp}
\alias{lift_shinyapp}
\title{Dockerize an Shiny App}
\usage{
lift_shinyapp(appDir = getwd(), appFiles = NULL,
  output_file = "docker.Rmd", output_dir = NULL, maintainer = NULL,
  email = NULL, shiny_base = "rocker/shiny")
}
\arguments{
\item{appDir}{Directory containing application. Defaults to current working directory.}

\item{appFiles}{The files and directories to bundle and deploy (only if upload = TRUE). Can be NULL, in which case all the files in the directory containing the application are bundled. Takes precedence over appFileManifest if both are supplied.}

\item{output_file}{A temporariy R markdown file with liftr header passed from shina app folder.}

\item{output_dir}{output_dir Directory to output \code{Dockerfile}. If not provided, will be the same directory as \code{input}.}

\item{maintainer}{maintainer information for Dockerfile}

\item{email}{email address for Dockerfile}

\item{shiny_base}{base image for shiny, by default it's rocker/shiny}
}
\description{
Parse dependecies from a shiny app folder and lift it into a Dockerfile
}
\examples{
\dontrun{
lift_shinayapp("test_app_folder")
}
}

