% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Onepunch-class.R
\docType{class}
\name{Onepunch-class}
\alias{Onepunch}
\alias{Onepunch-class}
\title{Creat a Onepunch object}
\arguments{
\item{input}{input Input (R Markdown or Shiny R markdown) file or shiny app folder.}

\item{output_dir}{output_dir Directory to output \code{Dockerfile}.}

\item{container}{container name}

\item{image}{image name}

\item{tag}{Docker image name to build, sent as docker argument \code{-t}.
If not specified, it will use the same name as the input file.}

\item{prebuild}{prebuild a command line string to call before docker build}

\item{build_args}{build_args A character string specifying additional
\code{docker build} arguments. For example,
\code{--pull=true -m="1024m" --memory-swap="-1"}.}

\item{liftr_template}{Rmarkdown template used to generate Dockerfile.}

\item{dockerfile}{Dockerfile path.}

\item{cache}{default TRUE, if FALSE, build with --no-cache=true}

\item{rm}{efault FALSE, if TRUE build with --rm}

\item{clean}{clean all container or not}

\item{type}{"shinyapp" or "shinydoc" or "rmd"}

\item{browseURL}{logical, default FALSE, to open browser automatically or not for shiny}

\item{shiny_run}{how to launch shiny from command line}

\item{url}{returned URL for browsing.}
}
\value{
Onepunch object
}
\description{
Onepunch object used for rendering
}
\section{Methods}{

\describe{
\item{\code{deploy(script = NULL, ...)}}{deploy inside container}
}}
\examples{
o = Onepunch("~/liftr_docker/ShinyDoc.Rmd")
\dontrun{
o$punch()
o$clean()
}
}

