% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift.R
\name{parse_rmd}
\alias{parse_rmd}
\title{parse Rmarkdown header}
\usage{
parse_rmd(input)
}
\description{
parse Rmarkdown header and return a list
}
\details{
The header section is use three hyphens --- as start line & end line,
or three hyphens --- as start line with three dots ...
as end line
}
\examples{
fl = system.file("examples/docker.Rmd", package = "liftr")
parse_rmd(fl)
}

