% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_all_auto}
\alias{prune_all_auto}
\title{Remove Docker Containers, Images, and Networks}
\usage{
prune_all_auto(volumes = FALSE)
}
\arguments{
\item{volumes}{Logical. Should we prune volumes? Default is \code{FALSE}.}
}
\value{
prune results
}
\description{
This function removes stopped containers, all networks
not used by at least one container, all dangling images,
and all build cache.
}
\examples{
\dontrun{
prune_all_auto()}
}
\references{
\url{https://docs.docker.com/engine/admin/pruning/}
}
