% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_container_auto}
\alias{prune_container_auto}
\title{Remove Dangling Docker Containers}
\usage{
prune_container_auto()
}
\value{
prune results
}
\description{
This function prunes all dangling Docker containers automatically.
}
\examples{
\dontrun{
prune_container_auto()}
}
\references{
\url{https://docs.docker.com/engine/admin/pruning/}
}
