% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune.R
\name{prune_image_auto}
\alias{prune_image_auto}
\title{Remove Dangling Docker Images}
\usage{
prune_image_auto(input_yml)
}
\value{
prune results
}
\description{
This function prunes all dangling Docker images automatically.
}
\examples{
\dontrun{
prune_image_auto()}
}
\references{
\url{https://docs.docker.com/engine/admin/pruning/}
}
