% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{purge_all}
\alias{purge_all}
\title{Purge Everything (Container + Image)}
\usage{
purge_all(container_name, image_name)
}
\arguments{
\item{container_name}{Name of the container to be stopped and removed.}

\item{image_name}{Name of the image to be removed.}
}
\value{
status
}
\description{
This function will try to purge the Docker container
and Docker image used for rendering the R Markdown document.
}
\examples{
\dontrun{
purge_all("liftr_container_uuid", "report")}
}

