% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{purge_container}
\alias{purge_container}
\title{Purge Docker Container}
\usage{
purge_container(container_name)
}
\arguments{
\item{container_name}{Name of the container to be stopped and removed.}
}
\value{
status
}
\description{
This function stops and removes the Docker container used
for rendering the R Markdown document, by running
\code{docker stop} and \code{docker rm} commands.
}
\examples{
\dontrun{
purge_container("liftr_container_uuid")}
}

