% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{purge_image}
\alias{purge_image}
\title{Purge Docker Image}
\usage{
purge_image(input)
}
\arguments{
\item{input}{The YAML file (default output of \code{\link{drender}})
storing the information of the image to be removed.}
}
\value{
status
}
\description{
This function removes the Docker image used
for rendering the R Markdown document by
running \code{docker rmi} commands.
}
\examples{
\dontrun{
purge_image("liftr-minimal.docker.yml")}
}

