% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{remove_all}
\alias{remove_all}
\title{Cleanup Everything}
\usage{
remove_all(container_name, image_name)
}
\arguments{
\item{container_name}{Name of the container to be stopped and removed.}

\item{image_name}{Name of the image to be removed.}
}
\value{
status
}
\description{
This function will try to stop and remove the Docker container
and image used for rendering the R Markdown document.
}
\examples{
\dontrun{
remove_all("liftr_container_uuid", "report")}
}

