% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{remove_container}
\alias{remove_container}
\title{Remove Docker Container}
\usage{
remove_container(container_name)
}
\arguments{
\item{container_name}{Name of the container to be removed.}
}
\value{
status
}
\description{
This function removes the Docker container used
for rendering the R Markdown document by
running \code{docker rm} commands.
}
\examples{
\dontrun{
remove_container("liftr_container_uuid")}
}

