% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{stop_container}
\alias{stop_container}
\title{Stop Docker Container}
\usage{
stop_container(container_name)
}
\arguments{
\item{container_name}{Name of the container to be stopped.}
}
\value{
status
}
\description{
This function stops the Docker container used
for rendering the R Markdown document by
running \code{docker stop} commands.
}
\examples{
\dontrun{
stop_container("liftr_container_uuid")}
}

