% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{addin_purge_image}
\alias{addin_purge_image}
\title{RStudio Addin: Purge Docker Image}
\usage{
addin_purge_image()
}
\description{
Call this function as an addin to removes the Docker image used
for rendering the current document.
}
\keyword{internal}

