% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/liftr-package.R
\docType{package}
\name{liftr-package}
\alias{liftr-package}
\title{liftr}
\description{
liftr
}
\details{
The vignette can be opened with \code{vignette('liftr')}.

\tabular{ll}{ Package: \tab liftr\cr Type: \tab Package\cr
License: \tab GPL\cr }
}
\author{
Nan Xiao <\email{nanx@uchicago.edu}>
        Teng-Fei Yin <\email{tengfei.yin@sbgenomics.com}>
}

