% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftr-package.R
\docType{package}
\name{liftr-package}
\alias{liftr-package}
\title{Containerize R Markdown Documents for Continuous Reproducibility}
\description{
Persistent reproducible reporting by
containerization of R Markdown documents.
}
\details{
\tabular{ll}{ Package: \tab liftr\cr Type: \tab Package\cr
License: \tab GPL-3\cr }
}
\author{
Nan Xiao <\email{me@nanx.me}>
Miao-Zhu Li <\email{miaozhu.li@duke.edu}>
Teng-Fei Yin <\email{tengfei.yin@sevenbridges.com}>
}
