% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftr-package.R
\docType{package}
\name{liftr-package}
\alias{liftr}
\alias{liftr-package}
\title{liftr: Containerize R Markdown Documents for Continuous Reproducibility}
\description{
Persistent reproducible reporting by containerization of R Markdown documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nanx.me/liftr/}
  \item \url{https://github.com/nanxstats/liftr}
  \item Report bugs at \url{https://github.com/nanxstats/liftr/issues}
}

}
\author{
\strong{Maintainer}: Nan Xiao \email{me@nanx.me} (\href{https://orcid.org/0000-0002-0250-5673}{ORCID})

Other contributors:
\itemize{
  \item Miao-Zhu Li \email{miaozhu.li@duke.edu} [contributor]
  \item Teng-Fei Yin \email{yintengfei@gmail.com} [contributor]
}

}
\keyword{internal}
