% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge.R
\name{purge_image}
\alias{purge_image}
\title{Purge Docker Image}
\usage{
purge_image(image_name)
}
\arguments{
\item{image_name}{Name of the image to be removed.}
}
\value{
status
}
\description{
This function removes the Docker image used
for rendering the R Markdown document by
running \code{docker rmi} commands.
}
\examples{
\dontrun{
purge_image("report")}
}

